Copy default configs to build directory. Finish python build scripts.

+++ prepare.py
@@ -4,6 +4,7 @@ import shutil
 import os
 import sys
 import subprocess
+import platform
 from typing import final
 
 
@@ -31,11 +32,14 @@ def download_dependencies():
                 print("`toml` package not installed! Aborting!")
                 exit(137)
 
-        # Download icons
+        # Check if icons exist already and download icons
         for icon in toml_data.get("icons"):
             try:
-                urllib.request.urlretrieve(ICONS_URL + icon + ".svg", ICONS_PATH + icon + ".svg")
-                print("Downloaded " + icon + ".svg")
+                if not os.path.exists(ICONS_PATH + icon + ".svg"):
+                    urllib.request.urlretrieve(ICONS_URL + icon + ".svg", ICONS_PATH + icon + ".svg")
+                    print("Downloaded " + icon + ".svg")
+                else:
+                    print("Skipping " + icon + ".svg, already exists!")
             except urllib.error.HTTPError:
                 print("Couldn't download " + icon + ".svg")
     # Other deps
@@ -44,6 +48,8 @@ def download_dependencies():
 def create_output():
     if not os.path.isdir("out"):
         os.mkdir("out")
+    if not os.path.isdir("out/bin"):
+        os.mkdir("out/bin")
 
 
 def check_buildtools():
@@ -59,13 +65,40 @@ def __check_buildtool(tool: str):
     print(tool + "... " + str(shutil.which(tool) is not None))
     return shutil.which(tool) is not None
 
+
 def build(release: bool):
-    print("Starting build tasks...")
-    release_str: str = ""
+    print("Starting build tasks (1/2)...")
+    release_flag: str = ""
+
+    # Launcher/Updater
     if release:
-        release_str = " --release"
-    exitcode = subprocess.call("cd ./Tiels && cargo build"+release_str+" && cp ./target/release/Tiels ../out",
+        release_flag = " --release"
+    exitcode = subprocess.call("cd ./Tiels && cargo build"+release_flag,
+                               shell=True)  # --out-dir is unstable.
+    # Exit script if it failed
+    if exitcode != 0:
+        print("Compilation error occurred! Aborting!")
+        exit(exitcode)
+    
+    # Copy compiled executable to output directory
+    print("Copying compiled binary...")
+    cp_ext: str = ""
+    cp_dr: str = "debug"
+    if platform.system() == "Windows":
+        cp_ext = ".exe"
+    if release:
+        cp_dr = "release"
+    shutil.copy("./Tiels/target/"+cp_dr+"/Tiels"+cp_ext, "./out/Tiels"+cp_ext)
+
+    # Main program
+    print("Starting build tasks (2/2)...")
+    release_flag = "build"
+    if release:
+        release_flag = "publish"
+    exitcode = subprocess.call("cd ./Avalonia.Tiels && dotnet "+release_flag+" -o ../out/bin",
                                shell=True)
+
+    # Exit script if it failed
     if exitcode != 0:
         print("Compilation error occurred! Aborting!")
         exit(exitcode)
